# map-business-backend/tests/contracts/openapi_snapshot.json
{
  "components": {
    "schemas": {
...
      "FeedbackAggregateRequest": {
        "properties": {
          "message_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "maxItems": 200,
            "minItems": 1,
            "title": "Message Ids",
            "type": "array"
          }
        },
        "required": [
          "message_ids"
        ],
        "title": "FeedbackAggregateRequest",
        "type": "object"
      },
      "FeedbackUpsertRequest": {
        "properties": {
          "correction_text": {
            "anyOf": [
              {
                "maxLength": 8000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Correction Text"
          },
          "rating": {
            "pattern": "^(helpful|unhelpful)$",
            "title": "Rating",
            "type": "string"
          },
          "reason_codes": {
            "items": {
              "type": "string"
            },
            "maxItems": 8,
            "title": "Reason Codes",
            "type": "array"
          },
          "reason_other": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason Other"
          }
        },
        "required": [
          "rating"
        ],
        "title": "FeedbackUpsertRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "MAP business management BFF. Frontend talks to this service only.",
    "title": "MAP Business Backend",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/admin/feedback": {
      "get": {
        "operationId": "admin_feedback_list_api_v1_admin_feedback_get",
        "parameters": [
          {
            "in": "query",
            "name": "rating",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^(helpful|unhelpful)$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Rating"
            }
          },
          {
            "in": "query",
            "name": "reason_code",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Reason Code"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Admin Feedback List Api V1 Admin Feedback Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin Feedback List"
      }
    },
    "/api/v1/admin/feedback/{feedback_id}:convert-to-evaluation-case": {
      "post": {
        "description": "R1-EVAL gate: conversion needs the evaluation draft/version fact source.\n\nR1-EVAL is not implemented in this round, so the capability is\nexplicitly unavailable behind a feature flag — never a fake case.",
        "operationId": "convert_to_evaluation_case_api_v1_admin_feedback__feedback_id__convert_to_evaluation_case_post",
        "parameters": [
          {
            "in": "path",
            "name": "feedback_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Feedback Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Convert To Evaluation Case Api V1 Admin Feedback  Feedback Id  Convert To Evaluation Case Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Convert To Evaluation Case"
      }
    },
...
    "/api/v1/conversations/{conversation_id}/feedback-summary": {
      "get": {
        "operationId": "conversation_feedback_summary_api_v1_conversations__conversation_id__feedback_summary_get",
        "parameters": [
          {
            "in": "path",
            "name": "conversation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Conversation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Conversation Feedback Summary Api V1 Conversations  Conversation Id  Feedback Summary Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Conversation Feedback Summary"
      }
    },
...
    "/api/v1/feedback/aggregate": {
      "post": {
        "description": "Counts for messages the caller may see; never other users' reasons.",
        "operationId": "aggregate_feedback_api_v1_feedback_aggregate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackAggregateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Aggregate Feedback Api V1 Feedback Aggregate Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Aggregate Feedback"
      }
    },
    "/api/v1/messages/{message_id}/feedback": {
      "delete": {
        "operationId": "delete_feedback_api_v1_messages__message_id__feedback_delete",
        "parameters": [
          {
            "in": "path",
            "name": "message_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Message Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Delete Feedback Api V1 Messages  Message Id  Feedback Delete",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Feedback"
      },
      "get": {
        "operationId": "get_feedback_api_v1_messages__message_id__feedback_get",
        "parameters": [
          {
            "in": "path",
            "name": "message_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Message Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "additionalProperties": true,
                      "type": "object"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Get Feedback Api V1 Messages  Message Id  Feedback Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Feedback"
      },
      "put": {
        "operationId": "upsert_feedback_api_v1_messages__message_id__feedback_put",
        "parameters": [
          {
            "in": "path",
            "name": "message_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Message Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackUpsertRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Upsert Feedback Api V1 Messages  Message Id  Feedback Put",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upsert Feedback"
      }
    },
    "/api/v1/messages/{message_id}/feedback/{kind}": {
      "delete": {
        "operationId": "delete_feedback_legacy_api_v1_messages__message_id__feedback__kind__delete",
        "parameters": [
          {
            "in": "path",
            "name": "message_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Message Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "kind",
            "required": true,
            "schema": {
              "title": "Kind",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Delete Feedback Legacy Api V1 Messages  Message Id  Feedback  Kind  Delete",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Feedback Legacy"
      }
    },
...
  }
}